+++ src/main/java/blackjack/domain/PlayerCards.java
@@ -1,5 +1,7 @@
-package blackjack.domain.card;
+package blackjack.domain;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
